requirement.txt
absl-py                   2.1.0
altair                    5.5.0
asttokens                 3.0.0
astunparse                1.6.3
attrs                     25.3.0
blinker                   1.9.0
cachetools                5.5.2
certifi                   2025.1.31
charset-normalizer        3.4.1
click                     8.2.1
colorama                  0.4.6
comm                      0.2.2
contourpy                 1.3.1
cycler                    0.12.1
debugpy                   1.8.12
decorator                 5.1.1
executing                 2.2.0
flatbuffers               25.1.24
fonttools                 4.56.0
gast                      0.6.0
gitdb                     4.0.12
GitPython                 3.1.44
google-pasta              0.2.0
greenlet                  3.2.1
grpcio                    1.70.0
h5py                      3.12.1
idna                      3.10
imbalanced-learn          0.13.0
ipykernel                 6.29.5
ipython                   8.32.0
jedi                      0.19.2
Jinja2                    3.1.6
joblib                    1.4.2
jsonschema                4.23.0
jsonschema-specifications 2025.4.1
jupyter_client            8.6.3
jupyter_core              5.7.2
keras                     3.8.0
kiwisolver                1.4.8
libclang                  18.1.1
Markdown                  3.7
markdown-it-py            3.0.0
MarkupSafe                3.0.2
matplotlib                3.10.0
matplotlib-inline         0.1.7
mdurl                     0.1.2
ml-dtypes                 0.4.1
namex                     0.0.8
narwhals                  1.40.0
nest-asyncio              1.6.0
numpy                     2.0.2
opencv-python             4.11.0.86
opt_einsum                3.4.0
optree                    0.14.0
packaging                 20.9
pandas                    2.2.3
parso                     0.8.4
pillow                    11.1.0
pip                       25.0
platformdirs              4.3.6
prompt_toolkit            3.0.50
protobuf                  5.29.3
psutil                    6.1.1
psycopg2                  2.9.10
pure_eval                 0.2.3
pyarrow                   20.0.0
pydeck                    0.9.1
Pygments                  2.19.1
pyparsing                 3.2.1
python-dateutil           2.9.0.post0
pytz                      2025.1
pywin32                   308
pyzmq                     26.2.1
referencing               0.36.2
requests                  2.32.3
rich                      13.9.4
rpds-py                   0.25.1
scikit-learn              1.6.1
scipy                     1.15.1
seaborn                   0.13.2
setuptools                75.8.0
six                       1.17.0
sklearn-compat            0.1.3
smmap                     5.0.2
SQLAlchemy                2.0.40
stack-data                0.6.3
streamlit                 1.45.1
tenacity                  9.1.2
tensorboard               2.18.0
tensorboard-data-server   0.7.2
tensorflow                2.18.0
tensorflow-hub            0.12.0
tensorflow_intel          2.18.0
tensorflowjs              3.18.0
termcolor                 2.5.0
threadpoolctl             3.6.0
toml                      0.10.2
tornado                   6.4.2
traitlets                 5.14.3
typing_extensions         4.12.2
tzdata                    2025.1
urllib3                   2.3.0
watchdog                  6.0.0
wcwidth                   0.2.13
Werkzeug                  3.1.3
wheel                     0.45.1
wrapt                     1.17.2
xgboost                   3.0.1